Added beginning of changes.
