+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/Extensions/APSExtension.cs
@@ -1,0 +1,16 @@
+namespace AdsByNimbus.Extensions
+{
+    #if NIMBUS_ENABLE_APS
+    public class APSExtension
+    {
+        public APSExtension(string appKey)
+        {
+#if UNITY_ANDROID
+            NimbusManager.Instance._configuration.androidApsAppKey = appKey;
+#elif UNITY_IOS
+            NimbusManager.Instance._configuration.iosApsAppKey = appKey;
+#endif
+        }
+    }
+    #endif
+}
