+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/Extensions/InMobiExtension.cs
@@ -1,0 +1,16 @@
+namespace AdsByNimbus.Extensions
+{
+    #if NIMBUS_ENABLE_INMOBI
+    public class InMobiExtension
+    {
+        public InMobiExtension(string accountId)
+        {
+#if UNITY_ANDROID
+            NimbusManager.Instance._configuration.androidInMobiAccountId = accountId;
+#elif UNITY_IOS
+            NimbusManager.Instance._configuration.iosInMobiAccountId = accountId;
+#endif
+        } 
+    }
+    #endif
+}
