+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/FullscreenAd.cs
@@ -1,5 +1,6 @@
 using AdsByNimbus;
 using AdsByNimbus.Internal;
+using AdsByNimbus.Internal.Utility;
 
 public class FullscreenAd: Ad
 {
@@ -7,7 +8,7 @@ public class FullscreenAd: Ad
     public float VideoBidFloor;
     
     public FullscreenAd(in AdEvents adEvents, string nimbusReportingPosition,
-        NimbusAdUnitPosition adPosition = NimbusAdUnitPosition.BOTTOM_CENTER, RequestModifiers? modifiers = null,
+        NimbusUnityPosition adPosition = NimbusUnityPosition.BOTTOM_CENTER, RequestModifiers? modifiers = null,
         float bannerBidFloor = 0f, float videoBidFloor = 0f) :
         base(AdType.Fullscreen, adEvents, nimbusReportingPosition, adPosition, modifiers)
     {
