+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/InlineAd.cs
@@ -5,14 +5,14 @@ using AdsByNimbus.Internal.Extensions;
 
 public class InlineAd: Ad
 {
-    public IabSupportedAdSizes BannerSize;
+    public AdSize BannerSize;
     public bool RespectSafeArea;
     public float BannerBidFloor;
     public int BannerRefreshIntervalInSeconds;
 
     public InlineAd(in AdEvents adEvents, string nimbusReportingPosition,
-        NimbusAdUnitPosition adPosition = NimbusAdUnitPosition.BOTTOM_CENTER, RequestModifiers? modifiers = null,
-        IabSupportedAdSizes adSize = IabSupportedAdSizes.Banner, bool respectSafeArea = false,
+        NimbusUnityPosition adPosition = NimbusUnityPosition.BOTTOM_CENTER, RequestModifiers? modifiers = null,
+        AdSize adSize = AdSize.banner, bool respectSafeArea = false,
         int bannerRefreshIntervalInSeconds = 30, float bannerBidFloor = 0f): 
         base(AdType.Inline, adEvents, nimbusReportingPosition, adPosition, modifiers)
     {
@@ -22,27 +22,27 @@ public class InlineAd: Ad
         BannerBidFloor = bannerBidFloor;
     }
 }
-public static class IabSupportedAdSizesExtension {
-    public static Tuple<int, int> ToWidthAndHeight(this IabSupportedAdSizes isa) {
+public static class AdSizesExtension {
+    public static Tuple<int, int> ToWidthAndHeight(this AdSize isa) {
         switch (isa) {
-            case IabSupportedAdSizes.Banner:
+            case AdSize.banner:
                 return new Tuple<int, int>(320, 50);
-            case IabSupportedAdSizes.FullScreenPortrait:
-                return new Tuple<int, int>(320, 480);
-            case IabSupportedAdSizes.FullScreenLandscape:
-                return new Tuple<int, int>(480, 320);
-            case IabSupportedAdSizes.HalfScreen:
-                return new Tuple<int, int>(300, 600);
-            case IabSupportedAdSizes.Letterbox:
+            case AdSize.mrec:
                 return new Tuple<int, int>(300, 250);
-            case IabSupportedAdSizes.LeaderBoard:
+            case AdSize.halfScreen:
+                return new Tuple<int, int>(300, 600);
+            case AdSize.leaderboard:
                 return new Tuple<int, int>(728, 90);
+            case AdSize.interstitialPortrait:
+                return new Tuple<int, int>(320, 480);
+            case AdSize.interstitialLandscape:
+                return new Tuple<int, int>(480, 320);
             default:
                 return new Tuple<int, int>(0, 0);
         }
     }
 }
-public enum NimbusAdUnitPosition
+public enum NimbusUnityPosition
 {
     BOTTOM_CENTER = 0,
     TOP_CENTER = 1,
@@ -53,11 +53,17 @@ public enum NimbusAdUnitPosition
     TOP_RIGHT = 6,
 }
 
-public enum IabSupportedAdSizes : byte {
-    Banner,
-    FullScreenPortrait,
-    FullScreenLandscape,
-    HalfScreen,
-    Letterbox,
-    LeaderBoard
+public enum AdSize : byte {
+    // Standard banner format (320×50)
+    banner,
+    // Medium rectangle (MREC) format (300×250)
+    mrec,
+    // Half-screen format (300×600)
+    halfScreen,
+    // Leaderboard format (728×90)
+    leaderboard,
+    // Interstitial portrait format (320×480)
+    interstitialPortrait,
+    // Interstitial landscape format (480×320)
+    interstitialLandscape,
 }
