+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/Internal/ConfigHelpers.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Runtime.InteropServices;
 using AOT;
 using AdsByNimbus.Internal.Extensions;
+using AdsByNimbus.RTB;
 using Newtonsoft.Json;
 using UnityEngine;
