--- com.adsbynimbus.nimbus/Runtime/AdsByNimbus/NimbusRTB.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 2dea2bfcfa394621b5589b8417f77e4e
-timeCreated: 1783010624
