+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/RTB/Request/environment.cs
@@ -1,0 +1,17 @@
+namespace AdsByNimbus.RTB
+{
+    /// <summary>
+    ///     Overrides the environment for a single ad.
+    /// </summary>
+    public struct environment
+    {
+        public string publisherKey; // Publisher key to be used for this ad
+        public string apiKey; // API Key to be used for this ad
+
+        public environment(string publisherKey, string apiKey)
+        {
+            this.publisherKey = publisherKey;
+            this.apiKey = apiKey;
+        }
+    }
+}
