+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/RequestComponent.cs
@@ -1,0 +1,7 @@
+namespace AdsByNimbus
+{
+    public interface RequestComponent
+    {
+        
+    }
+}
