+++ com.adsbynimbus.nimbus/Runtime/AdsByNimbus/RewardedAd.cs
@@ -7,7 +7,7 @@ public class RewardedAd: Ad
     public float VideoBidFloor;
 
     public RewardedAd(in AdEvents adEvents, string nimbusReportingPosition,
-        NimbusAdUnitPosition adPosition = NimbusAdUnitPosition.BOTTOM_CENTER, RequestModifiers? modifiers = null,
+        NimbusUnityPosition adPosition = NimbusUnityPosition.BOTTOM_CENTER, RequestModifiers? modifiers = null,
         float videoBidFloor = 0f) :
         base(AdType.Rewarded, adEvents, nimbusReportingPosition, adPosition, modifiers)
     {
