--- com.adsbynimbus.nimbus/Runtime/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 182bd7f9e2c0c426aaa5e7c321bc3d8b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
