--- com.adsbynimbus.nimbus/Runtime/Scripts/Nimbus.Internal.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9bc8de778d888449eab7c3308902cea3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
