--- com.adsbynimbus.nimbus/Runtime/Scripts/Nimbus.Internal/Extensions.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ad30d24033107475eb9a4d7042338fc5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
