--- com.adsbynimbus.nimbus/Runtime/Scripts/Nimbus.Internal/Interceptor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 509e242843b9e4d1081a67e9245a6f73
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
