--- com.adsbynimbus.nimbus/Runtime/Scripts/Nimbus.Internal/Interceptor/ThirdPartyDemand.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8af2e5f6c71394be2b14cd1c20a30476
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
